+++ src/lib.rs
@@ -26,6 +26,8 @@
 //!   * `tokio`: enable the asynchronous peer.
 //!   * `nalgebra`: implement conversions between `nalgebra` types and EGM messages.
 
+use std::time::Duration;
+
 mod error;
 pub use error::IncompleteTransmissionError;
 pub use error::ReceiveError;
@@ -48,6 +50,283 @@ pub mod tokio_peer;
 /// Conversions to/from nalgebra types.
 #[cfg(feature = "nalgebra")]
 mod nalgebra;
+
+impl msg::EgmHeader {
+	pub fn new(seqno: u32, timestamp_ms: u32, kind: msg::egm_header::MessageType) -> Self {
+		Self {
+			seqno: Some(seqno),
+			tm: Some(timestamp_ms),
+			mtype: Some(kind as i32),
+		}
+	}
+
+	/// Make a new command header.
+	pub fn command(seqno: u32, timestamp_ms: u32) -> Self {
+		Self::new(seqno, timestamp_ms, msg::egm_header::MessageType::MsgtypeCommand)
+	}
+
+	/// Make a new data header.
+	pub fn data(seqno: u32, timestamp_ms: u32) -> Self {
+		Self::new(seqno, timestamp_ms, msg::egm_header::MessageType::MsgtypeData)
+	}
+
+	/// Make a new correction header.
+	pub fn correction(seqno: u32, timestamp_ms: u32) -> Self {
+		Self::new(seqno, timestamp_ms, msg::egm_header::MessageType::MsgtypeCorrection)
+	}
+
+	/// Make a new path correction header.
+	pub fn path_correction(seqno: u32, timestamp_ms: u32) -> Self {
+		Self::new(seqno, timestamp_ms, msg::egm_header::MessageType::MsgtypePathCorrection)
+	}
+}
+
+impl msg::EgmCartesian {
+	/// Create a new cartesian position from x, y and z components.
+	pub fn new(x: f64, y: f64, z: f64) -> Self {
+		Self { x, y, z }
+	}
+
+	/// Get the cartesion position as [x, y, z] array.
+	pub fn as_array(&self) -> [f64; 3] {
+		[self.x, self.y, self.z]
+	}
+}
+
+impl From<[f64; 3]> for msg::EgmCartesian {
+	fn from(other: [f64; 3]) -> Self {
+		let [x, y, z] = other;
+		Self::new(x, y, z)
+	}
+}
+
+impl From<(f64, f64, f64)> for msg::EgmCartesian {
+	fn from(other: (f64, f64, f64)) -> Self {
+		let (x, y, z) = other;
+		Self::new(x, y, z)
+	}
+}
+
+impl msg::EgmQuaternion {
+	/// Create a new quaternion from w, x, y and z components.
+	pub fn from_wxyz(w: f64, x: f64, y: f64, z: f64) -> Self {
+		Self {
+			u0: w,
+			u1: x,
+			u2: y,
+			u3: z,
+		}
+	}
+
+	/// Get the quaternion as [w, x, y, z] array.
+	pub fn as_wxyz(&self) -> [f64; 4] {
+		[self.u0, self.u1, self.u2, self.u3]
+	}
+}
+
+impl msg::EgmEuler {
+	/// Create a new rotation from X, Y and Z rotations specified in degrees.
+	pub fn from_xyz_degrees(x: f64, y: f64, z: f64) -> Self {
+		Self { x, y, z }
+	}
+
+	/// Get the rotation as [x, y, z] array in degrees.
+	pub fn as_xyz_degrees(&self) -> [f64; 3] {
+		[self.x, self.y, self.z]
+	}
+}
+
+impl msg::EgmClock {
+	/// Create a new time point from seconds and microseconds.
+	pub fn new(sec: u64, usec: u64) -> Self {
+		Self { sec, usec }
+	}
+
+	/// Get the elapsed time since the epoch as [`Duration`].
+	///
+	/// Note that the duration will have only a microsecond resolution.
+	pub fn elapsed_since_epoch(&self) -> Duration {
+		Duration::new(self.sec, (self.usec) as u32 * 1000)
+	}
+}
+
+impl std::ops::Add<&Duration> for &msg::EgmClock {
+	type Output = msg::EgmClock;
+
+	fn add(self, right: &Duration) -> Self::Output {
+		let usec = self.usec + u64::from(right.subsec_micros());
+		msg::EgmClock {
+			sec: self.sec + right.as_secs() + usec / 1_000_000,
+			usec: usec % 1_000_000,
+		}
+	}
+}
+
+impl std::ops::Add<&msg::EgmClock> for &Duration {
+	type Output = msg::EgmClock;
+
+	fn add(self, right: &msg::EgmClock) -> Self::Output {
+		right + self
+	}
+}
+
+impl std::ops::Add<Duration> for msg::EgmClock {
+	type Output = Self;
+
+	fn add(self, right: Duration) -> Self::Output {
+		&self + &right
+	}
+}
+
+impl std::ops::Add<msg::EgmClock> for Duration {
+	type Output = msg::EgmClock;
+
+	fn add(self, right: msg::EgmClock) -> Self::Output {
+		&self + &right
+	}
+}
+
+impl std::ops::AddAssign<&Duration> for msg::EgmClock {
+	fn add_assign(&mut self, right: &Duration) {
+		*self = &*self + right
+	}
+}
+
+impl std::ops::AddAssign<Duration> for msg::EgmClock {
+	fn add_assign(&mut self, right: Duration) {
+		*self += &right
+	}
+}
+
+impl msg::EgmPose {
+	/// Create a new pose from a position and orientation.
+	pub fn new(position: impl Into<msg::EgmCartesian>, orientation: impl Into<msg::EgmQuaternion>) -> Self {
+		Self {
+			pos: Some(position.into()),
+			orient: Some(orientation.into()),
+			euler: None,
+		}
+	}
+}
+
+impl msg::EgmJoints {
+	/// Create a new joint list from a vector of joint values in degrees.
+	pub fn from_degrees(joints: impl Into<Vec<f64>>) -> Self {
+		Self { joints: joints.into() }
+	}
+}
+
+impl From<Vec<f64>> for msg::EgmJoints {
+	/// Create a new joint list from a vector of joint values in degrees.
+	fn from(other: Vec<f64>) -> Self {
+		Self::from_degrees(other)
+	}
+}
+
+impl From<&[f64]> for msg::EgmJoints {
+	/// Create a new joint list from a slice of joint values in degrees.
+	fn from(other: &[f64]) -> Self {
+		Self::from_degrees(other)
+	}
+}
+
+impl From<&[f64; 6]> for msg::EgmJoints {
+	/// Create a new joint list from an array of joint values in degrees.
+	fn from(other: &[f64; 6]) -> Self {
+		Self::from_degrees(*other)
+	}
+}
+
+impl msg::EgmExternalJoints {
+	/// Create a new external joint list from a vector of joint values in degrees.
+	pub fn from_degrees(joints: impl Into<Vec<f64>>) -> Self {
+		Self { joints: joints.into() }
+	}
+}
+
+impl From<Vec<f64>> for msg::EgmExternalJoints {
+	/// Create a new external joint list from a vector of joint values in degrees.
+	fn from(other: Vec<f64>) -> Self {
+		Self::from_degrees(other)
+	}
+}
+
+impl From<&[f64]> for msg::EgmExternalJoints {
+	/// Create a new external joint list from a slice of joint values in degrees.
+	fn from(other: &[f64]) -> Self {
+		Self::from_degrees(other)
+	}
+}
+
+impl msg::EgmPlanned {
+	/// Create a new joint target.
+	pub fn joints(joints: impl Into<msg::EgmJoints>, time: impl Into<msg::EgmClock>) -> Self {
+		Self {
+			time: Some(time.into()),
+			joints: Some(joints.into()),
+			cartesian: None,
+			external_joints: None,
+		}
+	}
+
+	/// Create a new cartesian target.
+	pub fn cartesian(pose: impl Into<msg::EgmPose>, time: impl Into<msg::EgmClock>) -> Self {
+		Self {
+			time: Some(time.into()),
+			cartesian: Some(pose.into()),
+			joints: None,
+			external_joints: None,
+		}
+	}
+}
+
+impl msg::EgmPathCorr {
+	/// Create a new path correction.
+	pub fn new(position: impl Into<msg::EgmCartesian>, age_ms: u32) -> Self {
+		Self {
+			pos: position.into(),
+			age: age_ms,
+		}
+	}
+}
+
+impl msg::EgmSensor {
+	/// Create a sensor message containing a joint target.
+	///
+	/// The header timestamp is created from the `time` parameter.
+	pub fn joint_target(sequence_number: u32, joints: impl Into<msg::EgmJoints>, time: impl Into<msg::EgmClock>) -> Self {
+		let time = time.into();
+		let timestamp_ms = (time.sec.wrapping_mul(1000) + time.usec / 1000) as u32;
+		Self {
+			header: Some(msg::EgmHeader::correction(sequence_number, timestamp_ms)),
+			planned: Some(msg::EgmPlanned::joints(joints, time)),
+			speed_ref: None,
+		}
+	}
+
+	/// Create a sensor message containing a cartesian target.
+	///
+	/// The header timestamp is created from the `time` parameter.
+	pub fn cartesian_target(sequence_number: u32, pose: impl Into<msg::EgmPose>, time: impl Into<msg::EgmClock>) -> Self {
+		let time = time.into();
+		let timestamp_ms = (time.sec.wrapping_mul(1000) + time.usec / 1000) as u32;
+		Self {
+			header: Some(msg::EgmHeader::correction(sequence_number, timestamp_ms)),
+			planned: Some(msg::EgmPlanned::cartesian(pose, time)),
+			speed_ref: None,
+		}
+	}
+}
+
+impl msg::EgmSensorPathCorr {
+	/// Create a sensor message containing a path correction.
+	pub fn new(sequence_number: u32, timestamp_ms: u32, correction: impl Into<msg::EgmCartesian>, age_ms: u32) -> Self {
+		Self {
+			header: Some(msg::EgmHeader::path_correction(sequence_number, timestamp_ms)),
+			path_corr: Some(msg::EgmPathCorr::new(correction, age_ms)),
+		}
+	}
+}
 
 impl msg::EgmRobot {
 	pub fn sequence_number(&self) -> Option<u32> {